Moved the dictionary into its own component, in order to make way for the flashcards
